TY-2524 fix potential 32bit problems (#240)
Does remove all usage of `usize`/`isize` in FFI.

Neither dart nor ffigen support that well when cross-compilation to 32bit targets is involved.

Once we have dart 1.16 we can replace this with a much better solutions using
`AbiSpecificInteger` and custom type-mapping options for `ffigen`.

**References:**

- [TY-2524](https://xainag.atlassian.net/browse/TY-2524)

+++ discovery_engine_core/providers/src/newscatcher.rs
@@ -64,7 +64,7 @@ pub struct Article {
 
     /// The page rank of the source website.
     #[serde(deserialize_with = "deserialize_rank")]
-    pub rank: usize,
+    pub rank: u64,
 
     /// The domain of the article's source, e.g. `xayn.com`. Not a valid URL.
     #[serde(
@@ -130,12 +130,12 @@ where
 }
 
 /// Null-value tolerant deserialization of rank
-fn deserialize_rank<'de, D>(deserializer: D) -> Result<usize, D::Error>
+fn deserialize_rank<'de, D>(deserializer: D) -> Result<u64, D::Error>
 where
     D: Deserializer<'de>,
 {
     let opt = Option::deserialize(deserializer)?;
-    Ok(opt.unwrap_or(usize::MAX))
+    Ok(opt.unwrap_or(u64::MAX))
 }
 
 /// Null-value tolerant deserialization of topic
